requirements.txt
rio-ui
qrcode[pil]
pillow
pyotp
numpy
pandas
plotly
matplotlib